Separate html, kss and template handling into different files
- Update kss-node to use sync parsing
