+++ package.json
@@ -23,7 +23,7 @@
     "handlebars": "^4.0.5",
     "html-loader": "^0.4.3",
     "html-webpack-plugin": "^2.21.0",
-    "kss": "^2.4.0",
+    "kss": "^3.0.0-beta.14"
     "memory-store": "0.0.1",
     "style-loader": "^0.13.1",
     "webpack": "^1.13.1",
